+++ src/wallet/test/util.h
@@ -5,11 +5,13 @@
 #ifndef BITCOIN_WALLET_TEST_UTIL_H
 #define BITCOIN_WALLET_TEST_UTIL_H
 
+#include <script/standard.h>
 #include <memory>
 
 class ArgsManager;
 class CChain;
 class CKey;
+enum class OutputType;
 namespace interfaces {
 class Chain;
 } // namespace interfaces
@@ -18,6 +20,12 @@ namespace wallet {
 class CWallet;
 
 std::unique_ptr<CWallet> CreateSyncedWallet(interfaces::Chain& chain, CChain& cchain, ArgsManager& args, const CKey& key);
+
+/** Returns a new encoded destination from the wallet (hardcoded to BECH32) */
+std::string getnewaddress(CWallet& w);
+/** Returns a new destination, of an specific type, from the wallet */
+CTxDestination getNewDestination(CWallet& w, OutputType output_type);
+
 } // namespace wallet
 
 #endif // BITCOIN_WALLET_TEST_UTIL_H
